Move panels to separate module
